third_party/blink/web_tests/http/tests/security/cross-frame-access-name-getter-expected.txt


----- tests for getting properties by name -----

PASS: canGet('targetWindow.testiframe') should be 'true' and is.
PASS: canGet('targetWindow.testimage') should be 'false' and is.
